requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cmdstanpy==0.9.68
convertdate==2.4.0
cycler==0.11.0
Cython==0.29.28
darts==0.18.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
ephem==4.1.3
executing==0.8.3
fonttools==4.31.2
frozenlist==1.3.0
fsspec==2022.3.0
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
hijri-converter==2.2.3
holidays==0.13
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.11.0
ipython==8.2.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.2.1
jupyter-core==4.9.2
kiwisolver==1.4.2
korean-lunar-calendar==0.2.1
lightgbm==3.3.2
LunarCalendar==0.0.9
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline==0.1.3
multidict==6.0.2
nest-asyncio==1.5.4
nfoursid==1.0.0
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.6.0
pmdarima==1.8.5
prompt-toolkit==3.0.28
prophet==1.0.1
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
Pygments==2.11.2
PyMeeus==0.5.11
pyparsing==3.0.7
pystan==2.19.1.1
python-dateutil==2.8.2
pytorch-lightning==1.6.0
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
setuptools-git==1.2
six==1.16.0
stack-data==0.2.0
statsmodels==0.13.2
tbats==1.1.0
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.11.0
torchmetrics==0.7.3
tornado==6.1
tqdm==4.63.1
traitlets==5.1.1
typing-extensions==4.1.1
ujson==5.1.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.0
xarray==2022.3.0
yarl==1.7.2
zipp==3.7.0
